+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ### Features
+* Create symbol states for trading systems [#61](https://github.com/Reiryoku-Technologies/Mida/pull/61)
 * Add `label` field to `MidaOrder` to represent the label indicated in the order directives [#63](https://github.com/Reiryoku-Technologies/Mida/pull/63)
 * Create a `clear()` method for clearing the internal state of a `MidaIndicator` [#62](https://github.com/Reiryoku-Technologies/Mida/pull/62)
 * Update documentation [#56](https://github.com/Reiryoku-Technologies/Mida/pull/56)